warning fix

+++ SurvivalZombie/classes/Zombie.h
@@ -14,7 +14,7 @@ public:
 	virtual void Action(b2Vec2 player_positon) ;
 private:
 	AIType* AIType;
-	double speed;
+	float32 speed;
 	sf::Sprite sprite;
 	int size; //rozmair zombie w pikselach
 };
